% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postInference.R
\name{posteriorInference}
\alias{posteriorInference}
\title{Posterior inference}
\usage{
posteriorInference(x, ...)
}
\arguments{
\item{x}{Bayesian Graphical Lasso object as returned by the function blockGLasso}
}
\description{
Posterior distributions of [LOH]
}
